+++ examples/bookshop/repos/catalog-service/src/main/java/demo/catalog/BookRepository.java
@@ -1,0 +1,5 @@
+package demo.catalog;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface BookRepository extends JpaRepository<Book, String> {}